% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_abonnement.R
\name{abonnement}
\alias{abonnement}
\title{Server-del av modul for abonnement i NoRGast sin shiny-app på Rapporteket}
\usage{
abonnement(input, output, session, reshID, userRole, hvd_session)
}
\value{
Modul abonnement
}
\description{
Kun til bruk i Shiny
}
