% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast_hj_funk.R
\name{abonnement_kvartal_norgast}
\alias{abonnement_kvartal_norgast}
\title{Generer kvartalsrapport og returner filnavn og sti til fil.}
\usage{
abonnement_kvartal_norgast(baseName, datoTil = Sys.Date(), reshID = 0,
  valgtShus = "")
}
\description{
Generer kvartalsrapport og returner filnavn og sti til fil.
}
