% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_abonnement.R
\name{abonnement_UI}
\alias{abonnement_UI}
\title{UI-del av modul for abonnement i NoRGast sin shiny-app på Rapporteket}
\usage{
abonnement_UI(id, BrValg)
}
\value{
Modul abonnement
}
\description{
Kun til bruk i Shiny
}
