% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{achocfigur2023rapport}
\alias{achocfigur2023rapport}
\title{Adhocfunksjon for årsrapportfigur}
\usage{
achocfigur2023rapport(
  valgtVar = "Malign",
  tittel = "Testtittel",
  Grvar1 = "Op_gr",
  Grvar2 = "Tilgang_utvidet",
  outfile = "",
  ...
)
}
\description{
Adhocfunksjon for årsrapportfigur
}
