% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_admtab.R
\name{admtab}
\alias{admtab}
\title{Serverdel av modul for Administrative tabeller-fane i NoRGast sin shiny-app på Rapporteket}
\usage{
admtab(
  input,
  output,
  session,
  reshID,
  RegData,
  userRole,
  hvd_session,
  skjemaoversikt,
  BrValg
)
}
\value{
Modulfunksjoner til Administrative tabeller
}
\description{
Kun til bruk i Shiny
}
