% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BrValgNorgastShiny.R
\name{BrValgNorgastShiny}
\alias{BrValgNorgastShiny}
\title{Funksjon som definerer brukervalgene til bruk i nedtrekksmenyer i Shiny app}
\usage{
BrValgNorgastShiny(RegData)
}
\value{
BrValg En liste bestående av brukervalgene
}
\description{
Funksjon som definerer brukervalgene til bruk i nedtrekksmenyer i Shiny app
}
