% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump}
\alias{datadump}
\title{Server-modul for datadump-fane i NoRGast sin shiny-app på Rapporteket}
\usage{
datadump(
  input,
  output,
  session,
  reshID,
  RegData,
  userRole,
  hvd_session,
  BrValg
)
}
\value{
Modulfunksjoner til datadump-fane
}
\description{
Kun til bruk i Shiny
}
