% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_server}
\alias{datadump_server}
\title{Server-modul for datadump-fane i NORGAST sin shiny-app på Rapporteket}
\usage{
datadump_server(id, reshID, RegData, userRole, brukerinfo, hvd_session, BrValg)
}
\value{
Modulfunksjoner til datadump-fane
}
\description{
Kun til bruk i Shiny
}
