% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datadump.R
\name{datadump_ui}
\alias{datadump_ui}
\title{UI-modul for datadump-fane i NORGAST sin shiny-app på Rapporteket}
\usage{
datadump_ui(id)
}
\value{
Modulfunksjoner til datadump-fane
}
\description{
Kun til bruk i Shiny
}
