% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datakvalitet.R
\name{datakval_server}
\alias{datakval_server}
\title{Server-modul for datakvalitetstabeller-fane i NORGAST sin shiny-app på Rapporteket}
\usage{
datakval_server(id, reshID, userRole, RegData, skjemaoversikt, hvd_session)
}
\value{
Modulfunksjoner til Datakvalitet
}
\description{
Kun til bruk i Shiny
}
