% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_datakvalitet.R
\name{datakval_ui}
\alias{datakval_ui}
\title{UI-modul for datakvalitetstabeller-fane i NoRGast sin shiny-app på Rapporteket}
\usage{
datakval_ui(id)
}
\value{
Modulfunksjoner til Datakvalitet
}
\description{
Kun til bruk i Shiny
}
