% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast_hj_funk.R
\name{dobbelreg}
\alias{dobbelreg}
\title{Identifiserer pasienter med flere enn ett forløp med samme operasjonsdato}
\usage{
dobbelreg(RegData)
}
\arguments{
\item{RegData}{Datasettet som kan inneholde dobbeltregistreringer}
}
\value{
En dataramme med utvalgte variabler for potensielt dobbeltregistrerte forløp
}
\description{
Identifiserer pasienter med flere enn ett forløp med samme operasjonsdato
}
