% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig}
\alias{fordelingsfig}
\title{Server-modul for fordelingsfigurer i NoRGast sin shiny-app på Rapporteket}
\usage{
fordelingsfig(input, output, session, reshID, RegData, userRole, hvd_session)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
