% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_server}
\alias{fordelingsfig_server}
\title{Server-modul for fordelingsfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
fordelingsfig_server(id, reshID, RegData, userRole, hvd_session, BrValg)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
