% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_fordelingsfig.R
\name{fordelingsfig_ui}
\alias{fordelingsfig_ui}
\title{UI-modul for fordelingsfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
fordelingsfig_ui(id)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
