% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_indikatorer.R
\name{indikatorfig}
\alias{indikatorfig}
\title{Server-modul for indikatorfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
indikatorfig(
  input,
  output,
  session,
  reshID,
  RegData,
  userRole,
  hvd_session,
  BrValg
)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
