% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_indikatorer.R
\name{indikatorfig_UI}
\alias{indikatorfig_UI}
\title{UI-modul for fordelingsfigurer i NoRGast sin shiny-app på Rapporteket}
\usage{
indikatorfig_UI(id)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
