% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast_hj_funk.R
\name{lag_kompl_tabell}
\alias{lag_kompl_tabell}
\title{Lag tabell over komplikasjoner til bruk i årsrapport}
\usage{
lag_kompl_tabell(regdata)
}
\description{
Lag tabell over komplikasjoner til bruk i årsrapport
}
