% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeStagingDataNoRGast.R, R/norgast_hj_funk.R
\name{makeStagingData}
\alias{makeStagingData}
\title{Make staging data for NoRGast}
\usage{
makeStagingData()

makeStagingData()
}
\value{
Character vector of staging files, invisibly

Character vector of staging files, invisibly
}
\description{
This function makes queries and pre-processing of registry data before
storing relevant staging data. Running this function may take a while so use
with care!

This function makes queries and pre-processing of registry data before
storing relevant staging data. Running this function may take a while so use
with care!
}
