% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast.R
\docType{package}
\name{norgast}
\alias{norgast}
\alias{norgast-package}
\title{All R resources needed for NoRGast on Rapporteket}
\description{
An implemetation of all R related resources for NoRGast on
Rapporteket, including R-functions for analysis and plotting,
noweb-files and even artwork needed for producing R-dependent reports
}
\section{Sample data}{

The package will soon contain a set of sample data that can be used for development
purposes. Not to be used in a PRODUCTION environment!
}

\section{Functions}{

Standard R-functions for processing registry data.
}

