% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{norgastAndelGruppert1Gr}
\alias{norgastAndelGruppert1Gr}
\title{Andeler gruppert på noe annet enn sykehusnavn}
\usage{
norgastAndelGruppert1Gr(
  valgtVar = "NyStomi",
  tittel = "Testtittel",
  Grvar1 = "AvstandAnalVerge_kat2",
  xtxt = "Andel (\%)",
  ytxt = "",
  outfile = "",
  ...
)
}
\description{
Andeler gruppert på noe annet enn sykehusnavn
}
