% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{norgastAndelGruppert2Gr_ggplot}
\alias{norgastAndelGruppert2Gr_ggplot}
\title{Adhocfunksjon for årsrapportfigur}
\usage{
norgastAndelGruppert2Gr_ggplot(
  valgtVar = "Malign",
  tittel = "Testtittel",
  Grvar1 = "Op_gr",
  Grvar2 = "Tilgang_utvidet",
  xtxt = "Andel (\%)",
  ytxt = "",
  outfile = "",
  ...
)
}
\description{
Adhocfunksjon for årsrapportfigur
}
