% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastAntallRegPrSykehus.R
\name{NorgastAntallRegPrSykehus}
\alias{NorgastAntallRegPrSykehus}
\title{Antall pasienter registrert per sykehus}
\usage{
NorgastAntallRegPrSykehus(RegData, datoFra = "2014-01-01",
  datoTil = "2050-12-31")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
Tabell En data frame over antall registreringer per sykehus, fordelt på ferdige og
uferdige hovedskjema og oppfølgingsskjema.
}
\description{
Denne funksjonen lager en liste over hvor mange pasienter som er registrert per sykehus.
}
