% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastApp.R
\name{norgastApp}
\alias{norgastApp}
\title{Run the NORGAST Shiny Application}
\usage{
norgastApp(logAsJson = TRUE, browser = FALSE)
}
\arguments{
\item{logAsJson}{Logical indicating whether to log messages as JSON}

\item{browser}{Logical indicating whether to open the app in a browser}
}
\value{
An object representing the NORGAST app
}
\description{
Run the NORGAST Shiny Application
}
