% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastBeregnIndikator.R
\name{norgastBeregnIndikator}
\alias{norgastBeregnIndikator}
\title{Beregn NORGAST sine indikatorer}
\usage{
norgastBeregnIndikator(RegData, ind_id)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{ind_id}{Indikatornavn}
}
\description{
Denne funksjonen beregner NORGAST sine offisielle kvalitetsindikatorer
}
