% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastFigAndelTid_kvartalsrapp.R
\name{NorgastFigAndelTid_kvartalsrapp}
\alias{NorgastFigAndelTid_kvartalsrapp}
\title{Tidstrend (år/mnd) av rate/andel for en gitt variabel}
\usage{
NorgastFigAndelTid_kvartalsrapp(RegData = 0, valgtVar = "ReLapNarkose",
  datoFra = "2014-01-01", datoTil = "2050-12-31", minald = 0,
  maxald = 130, erMann = 99, outfile = "", reshID,
  enhetsUtvalg = 1, preprosess = F, inkl_konf = 99, malign = 99,
  elektiv = 99, BMI = "", tilgang = 99, valgtShus = c(""),
  minPRS = 0, maxPRS = 2, ASA = "", whoEcog = "",
  forbehandling = 99, hentData = F, tidsenhet = "Aar",
  reseksjonsGr = "", ncsp = "")
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}

\item{inkl_konf}{Inkluder konfidensintervall i figur
0: Nei
1: Ja
99: Bruk default som definert i NorgastPrepVar}

\item{malign}{Er diagnosen malign eller benign
0: Benign
1: Malign
99: Alle}

\item{elektiv}{Elektiv eller øyeblikkelig hjelp
0: Øyeblikkelig hjelp
1: Elektiv
99: Begge deler (Default)}

\item{BMI}{BMI-klasse, flervalg hvor (Default alle)
1: Alvorlig undervekt
2: Undervekt
3: Mild undervekt
4: Normal
5: Overvekt
6: Moderat fedme, klasse I
7: Fedme, klasse II
8: Fedme, klasse III}

\item{tilgang}{Tilgang i abdomen
1: Åpen eller konvertert
2: Lapaoskopisk
99: Alle (Default)}

\item{valgtShus}{Vektor med AvdResh over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{minPRS}{Minimum PRS (Default 0?)}

\item{maxPRS}{Maksimum PRS (Default 2?)}

\item{ASA}{ASA-grad, flervalg hvor (Default alle)
1: Ingen organisk, fysiologisk, biokjemisk eller psykisk forstyrrelse.
Den aktuelle lidelsen er lokalisert og gir ikke generelle systemforstyrrelser.
2: Moderat sykdom eller forstyrrelser som ikke forårsaker funksjonelle begrensninger.
3: Alvorlig sykdom eller forstyrrelse som gir definerte funksjonelle begrensninger.
4: Livstruende organisk sykdom som ikke behøver å være knyttet til den aktuelle
kirurgiske lidelsen eller som ikke alltid bedres ved det planlagte kirurgiske inngrepet.
5: Døende pasient som ikke forventes å overleve 24 timer uten kirurgi.}

\item{whoEcog}{WHO WCOG score, flervalg hvor (Default alle)
0: Fullt aktiv
1: Lett husarbeid og sittende arbeid
2: Oppe > 50% av dagen, selvstelt
3: Oppe < 50% av dagen, delvis selvstelt
4: Kun i stol/seng, hjelp til alt stell
9: Ukjent}

\item{forbehandling}{Onkologisk forbehandling
1: Cytostatika
2: Stråleterapi
3: Komb. kjemo/radioterapi
4: Ingen
99: Alle}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}

\item{tidsenhet}{Plot figur med år eller måned som tidsenhet
'Aar' (Default)
'Mnd'}

\item{ncsp}{NCSP-koder(r) som skal være inkludert i utvalget}
}
\value{
En figur med tidsutvikling av rate over år
}
\description{
Årlige eller månedlige rater for valgt variabel.
Konfidensintervall kan inkluderes hvis ønskelig.
}
\details{
Konfidensintervallet er basert på Clopper Pearsons "eksakte" metode for binominalfordelt data.
}
