% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastFigAndelTid_SammenligUtvalg.R
\name{NorgastFigAndelTid_SammenligUtvalg}
\alias{NorgastFigAndelTid_SammenligUtvalg}
\title{Tidstrend (år/kvartal/mnd) av rate/andel for en gitt variabel for to utvalg}
\usage{
NorgastFigAndelTid_SammenligUtvalg(
  plotdata = 0,
  outfile = "",
  tidsenhet = "Aar",
  inkl_konf = FALSE,
  datoFra = "2014-01-01",
  datoTil = today(),
  fra0 = FALSE,
  inkl_tall = FALSE
)
}
\arguments{
\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{tidsenhet}{Plot figur med år eller måned som tidsenhet
'Aar' (Default)
'Kvartal'
'Halvaar'
'Mnd'}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}
}
\value{
En figur med tidsutvikling av rate over år for to utvalg
}
\description{
Årlige eller månedlige rater for valgt variabel.
Konfidensintervall kan inkluderes hvis ønskelig.
}
\details{
Konfidensintervallet er basert på Clopper Pearsons "eksakte" metode for binominalfordelt data.
}
