% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastFigAntRegTid.R
\name{norgastFigAntRegTid}
\alias{norgastFigAntRegTid}
\title{Plot andeler/andeler i angitt format}
\usage{
norgastFigAntRegTid(RegData, tittel = "", width = 800, height = 700,
  sideTxt = "Boområde/opptaksområde", decreasing = F, terskel = 30,
  minstekrav = NA, maal = NA, skriftStr = 1.3, pktStr = 1.4,
  legPlass = "top", minstekravTxt = "Min.", maalTxt = "Mål",
  graaUt = NA, inkl_konf = F, datoFra = "2014-01-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = 99,
  outfile = "", preprosess = F, malign = 99, elektiv = 99,
  BMI = "", tilgang = 99, minPRS = 0, maxPRS = 2, ASA = "",
  whoEcog = "", forbehandling = 99, hentData = 0,
  reseksjonsGr = "", ncsp = "")
}
\arguments{
\item{graaUt}{En vektor med navn på enheter som skal ha grå søyler}

\item{outfile}{Angir filnavn og format på figuren som returneres,}

\item{AntTilfeller}{En dataramme med antall i spesifisert form}

\item{N}{En dataramme med nevneren i andelsberegningen}
}
\value{
Et plot av andeler over tre år
}
\description{
Denne funksjonen tar som input en dataramme med andeler over 3 år,
der radnavn angir grupperingsvariabel og kolonnenavn år. Funksjonen
returnerer et søyleplot hvor søylene representerer sist år, fyllt sirkel er året
før og åpen sirkel to år før
}
