% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{norgastFordelingOpGruppert}
\alias{norgastFordelingOpGruppert}
\title{Fordeling av operasjoner over valgt variabel
og splittet på grupperingsvariabel}
\usage{
norgastFordelingOpGruppert(
  outfile = "",
  Grvar1 = "AvstandAnalVerge_fakt",
  Grvar2 = "Tilgang_utvidet",
  tittel = c("Fordeling av operasjoner", "over avstand fra analkanten"),
  ytxt = "Andel (\%)",
  xtxt = "Avstand analkanten (cm)",
  rev_farger = FALSE,
  ...
)
}
\description{
Fordeling av operasjoner over valgt variabel
og splittet på grupperingsvariabel
}
