% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastHentSkjemaOversikt.R
\name{NorgastHentSkjemaOversikt}
\alias{NorgastHentSkjemaOversikt}
\title{Provide SkjemaOversikt for NoRGast}
\usage{
NorgastHentSkjemaOversikt()
}
\value{
SkjemaOversikt data frame
}
\description{
Provides SkjemaOversikt for NoRGast data staging
}
