% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastBeregnIndikator.R
\name{norgastPlotIndikator}
\alias{norgastPlotIndikator}
\title{Plot NORGAST sine indikatorer}
\usage{
norgastPlotIndikator(
  Indikator,
  tittel = "",
  decreasing = FALSE,
  terskel = 10,
  lavDG = "",
  lavDGtekst = "Dekningsgrad < 60 \%",
  width = 800,
  height = 700,
  outfile = "",
  graaUt = NA,
  skriftStr = 1.3,
  utvalgTxt = "",
  minstekrav = NA,
  maal = NA,
  pktStr = 1.4,
  legPlass = "top",
  minstekravTxt = "Akseptabelt",
  maalTxt = "Mål",
  maalretn = "hoy",
  prikktall = TRUE
)
}
\arguments{
\item{Indikator}{indikatordata som beregnet i norgastBeregnIndikator}

\item{tittel}{Tittel på plot}
}
\description{
Denne funksjonen plotter NORGAST sine offisielle kvalitetsindikatorer
basert på beregningene i norgastBeregnIndikator
}
