% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastPrepVar.R
\name{NorgastPrepVar}
\alias{NorgastPrepVar}
\title{Preparer variabler for plotting}
\usage{
NorgastPrepVar(RegData, valgtVar, enhetsUtvalg = 1)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}
}
\value{
PrepData En liste med plotrelevante størrelser
}
\description{
Denne funksjonen grupperer og klargjør variabler for andelsplot
}
\details{
Her kan detaljer skrives
}

