% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastTabeller.R
\name{NorgastTabeller}
\alias{NorgastTabeller}
\title{Lag en tabell over de vanligste operasjonstypene samt en tabell med de komplikasjonsrater for de ulike
komplikasjonstypene for de obligatoriske operasjonsgruppene.}
\usage{
NorgastTabeller(RegData = RegData, datoFra = "2014-01-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = 99,
  enhetsUtvalg = 0, Terskel = 15, reshID = reshID)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{Terskel}{- Minste antall registreringer av operasjonstype for at den skal telles med i lista
over operasjonstyper}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}
}
\value{
Tabell En tabell med de vanligste operasjonstypene
        Tabell2 En tabell med komplikasjonsrater for de ulike komplikasjonstypene
        Terskel Minste antall
}
\description{
Lag en tabell over de vanligste operasjonstypene samt en tabell med de komplikasjonsrater for de ulike
komplikasjonstypene for de obligatoriske operasjonsgruppene.
}

