% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_overlevelse.R
\name{overlevelse}
\alias{overlevelse}
\title{Server-modul for overlevelesesfigurer i NoRGast sin shiny-app på Rapporteket}
\usage{
overlevelse(
  input,
  output,
  session,
  reshID,
  RegData,
  userRole,
  hvd_session,
  BrValg
)
}
\value{
Modul overlevelesesfigurer
}
\description{
Kun til bruk i Shiny
}
