% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_overlevelse.R
\name{overlevelse_server}
\alias{overlevelse_server}
\title{Server-modul for overlevelesesfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
overlevelse_server(id, reshID, RegData, userRole, hvd_session, BrValg)
}
\value{
Modul overlevelesesfigurer
}
\description{
Kun til bruk i Shiny
}
