% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sammenlign_utvalg_tid.R
\name{saml_andeler}
\alias{saml_andeler}
\title{Server-modul for sammenligning av andeler i NoRGast sin shiny-app på Rapporteket}
\usage{
saml_andeler(input, output, session, reshID, RegData, userRole, hvd_session)
}
\value{
Modul sammenligning av andeler
}
\description{
Kun til bruk i Shiny
}
