% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_samledok.R
\name{samledok}
\alias{samledok}
\title{Server-modul for Samledokumenter-fane i NoRGast sin shiny-app på Rapporteket}
\usage{
samledok(
  input,
  output,
  session,
  reshID,
  RegData,
  userRole,
  hvd_session,
  BrValg
)
}
\value{
Modulfunksjoner til Samledokumenter
}
\description{
Kun til bruk i Shiny
}
