% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_samledok.R
\name{samledok_ui}
\alias{samledok_ui}
\title{UI-modul for Samledokumenter-fane i NORGAST sin shiny-app på Rapporteket}
\usage{
samledok_ui(id)
}
\value{
Modulfunksjoner til Samledokumenter
}
\description{
Kun til bruk i Shiny
}
