% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sykehusvisning.R
\name{sykehusvisning}
\alias{sykehusvisning}
\title{Server-modul for sykehusvise andeler i NoRGast sin shiny-app på Rapporteket}
\usage{
sykehusvisning(input, output, session, reshID, RegData, hvd_session, BrValg)
}
\value{
Modul sykehusvisning, andeler
}
\description{
Kun til bruk i Shiny
}
