% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sykehusvisning.R
\name{sykehusvisning_server}
\alias{sykehusvisning_server}
\title{Server-modul for sykehusvise andeler i NORGAST sin shiny-app på Rapporteket}
\usage{
sykehusvisning_server(id, RegData, hvd_session, BrValg)
}
\value{
Modul sykehusvisning, andeler
}
\description{
Kun til bruk i Shiny
}
