% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sykehusvisning.R
\name{sykehusvisning_ui}
\alias{sykehusvisning_ui}
\title{UI-modul for sykehusvise andeler i NORGAST sin shiny-app på Rapporteket}
\usage{
sykehusvisning_ui(id)
}
\value{
Modul sykehusvisning, andeler
}
\description{
Kun til bruk i Shiny
}
