% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_tidsvisning.R
\name{tidsvisning_server}
\alias{tidsvisning_server}
\title{Server-modul for tidsvisning andeler i NORGAST sin shiny-app på Rapporteket}
\usage{
tidsvisning_server(id, reshID, RegData, userRole, hvd_session, BrValg)
}
\value{
Modul tidsvisning, andeler
}
\description{
Kun til bruk i Shiny
}
