% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_tidsvisning.R
\name{tidsvisning_ui}
\alias{tidsvisning_ui}
\title{UI-modul for tidsvisning andeler i NORGAST sin shiny-app på Rapporteket}
\usage{
tidsvisning_ui(id)
}
\value{
Modul tidsvisning, andeler
}
\description{
Kun til bruk i Shiny
}
