% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_traktplot.R
\name{traktplot}
\alias{traktplot}
\title{UI-modul for sykehusvise andeler sammen med traktplot i NoRGast sin shiny-app på Rapporteket}
\usage{
traktplot(input, output, session, reshID, RegData, hvd_session, BrValg)
}
\value{
Modul traktplot
}
\description{
Kun til bruk i Shiny
}
