% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_traktplot.R
\name{traktplot_server}
\alias{traktplot_server}
\title{UI-modul for sykehusvise andeler sammen med traktplot i NORGAST sin shiny-app på Rapporteket}
\usage{
traktplot_server(id, reshID, RegData, hvd_session, BrValg)
}
\value{
Modul traktplot
}
\description{
Kun til bruk i Shiny
}
