% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_traktplot.R
\name{traktplot_UI}
\alias{traktplot_UI}
\title{UI-modul for sykehusvise andeler sammen med traktplot i NoRGast sin shiny-app på Rapporteket}
\usage{
traktplot_UI(id)
}
\value{
Modul traktplot
}
\description{
Kun til bruk i Shiny
}
