% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast_hj_funk.R
\name{write.csv3}
\alias{write.csv3}
\title{Skriv til csv i nordisk format med Latin1 som default tegnsetting}
\usage{
write.csv3(x, file = "", tegnsetting = "latin1", ...)
}
\description{
Skriv til csv i nordisk format med Latin1 som default tegnsetting
}
