% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgast_hj_funk.R
\name{abonnement_kvartal_norgast}
\alias{abonnement_kvartal_norgast}
\title{Generer kvartalsrapport og returner filnavn og sti til fil.}
\usage{
abonnement_kvartal_norgast(
  baseName,
  reshID = 0,
  valgtShus = "",
  brukernavn = "Pjotr"
)
}
\description{
Generer kvartalsrapport og returner filnavn og sti til fil.
}
