% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{achocfigur2023rapport}
\alias{achocfigur2023rapport}
\title{Adhocfunksjon for årsrapportfigur}
\usage{
achocfigur2023rapport(
  RegData,
  datoFra = "2023-01-01",
  datoTil = "2023-12-31",
  outfile = ""
)
}
\description{
Adhocfunksjon for årsrapportfigur
}
