% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastFigAndeler.R
\name{FigAndeler}
\alias{FigAndeler}
\title{Lag søylediagram eller stabelplott som viser andeler av ulike variabler}
\usage{
FigAndeler(RegData = 0, valgtVar = "Alder", datoFra = "2014-01-01",
  datoTil = "2050-12-31", minald = 0, maxald = 130, erMann = 99,
  op_gruppe = 0, outfile = "", reshID, enhetsUtvalg = 1, stabel = F,
  andel = T, preprosess = T, elektiv = 99, BMI = "",
  valgtShus = c(""), hentData = T)
}
\arguments{
\item{RegData}{En dataramme med alle nødvendige variabler fra registeret}

\item{valgtVar}{Hvilken variabel skal plottes}

\item{datoFra}{Tidligste dato i utvalget (vises alltid i figuren).}

\item{datoTil}{Seneste dato i utvalget (vises alltid i figuren).}

\item{minald}{Alder, fra og med (Default: 0)}

\item{maxald}{Alder, til og med (Default: 130)}

\item{erMann}{kjønn
1: menn
0: kvinner
99: begge (alt annet enn 0 og 1) (Default)}

\item{op_gruppe}{Operasjonsgruppe
0: Alle grupper (Default)
1: Kolonreseksjoner
2: Rektumreseksjoner
3: Øsofagusreseksjoner
4: Ventrikkelreseksjoner
5: Leverreseksjoner
6: Whipple's operasjon
9: Annet}

\item{outfile}{Navn på fil figuren skrives til. Default: '' (Figur skrives
til systemets default output device (som regel skjerm))}

\item{reshID}{Parameter følger fra innlogging helseregister.no og angir
hvilken enhet i spesialisthelsetjenesten brukeren tilhører}

\item{enhetsUtvalg}{Lag figur for
0: Hele landet
1: Egen enhet mot resten av landet (Default)
2: Egen enhet}

\item{stabel}{Stabelplot
0: Stabel
1: Søyle (Default)}

\item{andel}{Plot antall eller andeler
TRUE: andeler (Default)
FALSE: antall}

\item{preprosess}{Preprosesser data
FALSE: Nei (Default)
TRUE: Ja}

\item{elektiv}{Elektiv eller øyeblikkelig hjelp
0: Øyeblikkelig hjelp
1: Elektiv
99: Begge deler (Default)}

\item{BMI}{BMI-klasse}

\item{valgtShus}{Vektor med AvdResh over hvilke sykehus man genererer rapporten for.
Denne overstyrer reshID og er bare tilgjengelig for SC-bruker.}

\item{hentData}{Gjør spørring mot database
FALSE: Nei, RegData gis som input til funksjonen (Default)
TRUE: Ja}
}
\value{
En figur med søylediagram eller et stabelplot av ønsket variabel
}
\description{
Denne funksjonen lager et søylediagram eller stabelplot som viser andeler av valgt variabel
filtrert på de utvalg som er gjort.
}

