% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_indikatorer.R
\name{indikatorfig_ui}
\alias{indikatorfig_ui}
\title{UI-modul for indikatorfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
indikatorfig_ui(id)
}
\value{
Modul fordelingsfigur
}
\description{
Kun til bruk i Shiny
}
