% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{norgastAndelGruppert1Gr}
\alias{norgastAndelGruppert1Gr}
\title{Andeler gruppert på noe annet enn sykehusnavn}
\usage{
norgastAndelGruppert1Gr(
  valgtVar = "Anastomoselekkasje",
  tittel = "Testtittel",
  Grvar1 = "Robot",
  xtxt = "Andel (\%)",
  ytxt = "",
  inkl_konf = TRUE,
  outfile = "",
  ...
)
}
\description{
Andeler gruppert på noe annet enn sykehusnavn
}
