% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/achocfigur2023rapport.R
\name{norgastAndelGruppert2Gr}
\alias{norgastAndelGruppert2Gr}
\title{Adhocfunksjon for årsrapportfigur}
\usage{
norgastAndelGruppert2Gr(
  valgtVar = "Malign",
  tittel = "Testtittel",
  Grvar1 = "Op_gr",
  Grvar2 = "Tilgang_utvidet",
  xtxt = "Andel (\%)",
  ytxt = "",
  inkl_konf = FALSE,
  pst_kol = TRUE,
  outfile = "",
  farger = NA,
  ...
)
}
\description{
Adhocfunksjon for årsrapportfigur
}
