% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastApp.R
\name{norgastApp}
\alias{norgastApp}
\title{Run the NORGAST Shiny Application}
\usage{
norgastApp(logAsJson = TRUE)
}
\value{
An object representing the NORGAST app
}
\description{
Run the NORGAST Shiny Application
}
