% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NorgastHentKolnavn.R
\name{NorgastHentKolnavn}
\alias{NorgastHentKolnavn}
\title{Hent navnene til variablene i datadump NoRGast}
\usage{
NorgastHentKolnavn(tabellnavn = "AlleVarNum")
}
\value{
kolnavn dataframe
}
\description{
Hent navnene til variablene i datadump NoRGast
}
