% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norgastBeregnIndikator.R
\name{norgastPlotIndikator}
\alias{norgastPlotIndikator}
\title{Plot NORGAST sine indikatorer}
\usage{
norgastPlotIndikator(
  AntTilfeller,
  N,
  andeler,
  tittel = "",
  decreasing = FALSE,
  terskel = 10,
  lavDG = "",
  lavDGtekst = "Dekningsgrad < 60 \%",
  width = 600,
  height = 700,
  outfile = "",
  graaUt = NA,
  skriftStr = 1.2,
  utvalgTxt = "",
  minstekrav = NA,
  maal = NA,
  pktStr = 1.4,
  legPlass = "top",
  minstekravTxt = "Akseptabelt",
  maalTxt = "M<U+00E5>l",
  maalretn = "hoy",
  prikktall = TRUE,
  pst_kolonne = T
)
}
\arguments{
\item{tittel}{Tittel på plot}

\item{Indikator}{indikatordata som beregnet i norgastBeregnIndikator}
}
\description{
Denne funksjonen plotter NORGAST sine offisielle kvalitetsindikatorer
basert på beregningene i norgastBeregnIndikator
}
