% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_overlevelse.R
\name{overlevelse_UI}
\alias{overlevelse_UI}
\title{UI-modul for overlevelesesfigurer i NORGAST sin shiny-app på Rapporteket}
\usage{
overlevelse_UI(id)
}
\value{
Modul overlevelesesfigurer
}
\description{
Kun til bruk i Shiny
}
