% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sammenlign_utvalg_tid.R
\name{saml_andeler_UI}
\alias{saml_andeler_UI}
\title{UI-modul for sammenligning av andeler i NORGAST sin shiny-app på Rapporteket}
\usage{
saml_andeler_UI(id)
}
\value{
Modul sammenligning av andeler
}
\description{
Kun til bruk i Shiny
}
