% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_samledok.R
\name{samledok_UI}
\alias{samledok_UI}
\title{UI-modul for Samledokumenter-fane i NoRGast sin shiny-app på Rapporteket}
\usage{
samledok_UI(id, BrValg)
}
\value{
Modulfunksjoner til Samledokumenter
}
\description{
Kun til bruk i Shiny
}
