% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_startside.R
\name{startside}
\alias{startside}
\title{Server-del av startside for NORGAST sin Rapporteket-app}
\usage{
startside(input, output, session, usrRole)
}
\description{
Server-del av startside for NORGAST sin Rapporteket-app
}
