% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_sykehusvisning.R
\name{sykehusvisning_UI}
\alias{sykehusvisning_UI}
\title{UI-modul for sykehusvise andeler i NoRGast sin shiny-app på Rapporteket}
\usage{
sykehusvisning_UI(id, BrValg)
}
\value{
Modul sykehusvisning, andeler
}
\description{
Kun til bruk i Shiny
}
