% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_tidsvisning.R
\name{tidsvisning_UI}
\alias{tidsvisning_UI}
\title{UI-modul for tidsvisning andeler i NoRGast sin shiny-app på Rapporteket}
\usage{
tidsvisning_UI(id, BrValg)
}
\value{
Modul tidsvisning, andeler
}
\description{
Kun til bruk i Shiny
}
