% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modul_traktplot.R
\name{traktplot_ui}
\alias{traktplot_ui}
\title{UI-modul for sykehusvise andeler sammen med traktplot i NORGAST sin shiny-app på Rapporteket}
\usage{
traktplot_ui(id)
}
\value{
Modul traktplot
}
\description{
Kun til bruk i Shiny
}
